+++ drizzle.config.ts
@@ -9,5 +9,9 @@ export default {
     dbCredentials: {
         url: process.env.DATABASE_URL!,
         authToken: process.env.DATABASE_AUTH_TOKEN!
-    }
+    },
+    // Print all statements
+    verbose: true,
+    // Always ask for my confirmation
+    strict: true,
 } satisfies Config;