+++ src/app.d.ts
@@ -1,6 +1,8 @@
 // See https://kit.svelte.dev/docs/types#app
 // for information about these interfaces
 /// <reference types="svelte-adapter-azure-swa" />
+// Do not add any other lines of code to this file!
+import "@total-typescript/ts-reset";
 
 declare global {
 	namespace App {