+++ src/lib/components/ui/dropdown-menu/dropdown-menu-content.svelte
@@ -1,0 +1,21 @@
+<script lang="ts">
+	import { DropdownMenu as DropdownMenuPrimitive } from 'bits-ui';
+	import { cn, flyAndScale } from '$lib/helpers/utils';
+
+	type $$Props = DropdownMenuPrimitive.ContentProps;
+
+	let className: $$Props['class'] = undefined;
+	export let transition: $$Props['transition'] = flyAndScale;
+	export let transitionConfig: $$Props['transitionConfig'] = undefined;
+	export { className as class };
+</script>
+
+<DropdownMenuPrimitive.Content
+	{transition}
+	{transitionConfig}
+	class={cn('z-50 min-w-[8rem] border bg-white border-slate-500 p-1 focus:outline-none', className)}
+	{...$$restProps}
+	on:keydown
+>
+	<slot />
+</DropdownMenuPrimitive.Content>