+++ src/routes/(auth)/google/callback/+server.ts
@@ -1,7 +1,10 @@
+import { isOsteopath } from '$lib/helpers/utils.js';
+import { db } from '$lib/server/db/index.js';
+import { osteopath as osteopathTable } from '$lib/server/db/schema/index.js';
 import { auth, googleAuth } from '$lib/server/lucia.js';
 import { OAuthRequestError } from '@lucia-auth/oauth';
 
-export const GET = async ({ url, cookies, locals, request }) => {
+export const GET = async ({ url, cookies, locals }) => {
   console.log("Response From Callback")
   const session = await locals.auth.validate();
   if (session) {
@@ -36,6 +39,17 @@ export const GET = async ({ url, cookies, locals, request }) => {
           phone_number: null
         }
       });
+      if( isOsteopath(googleUser.email) ) {
+        try {
+          console.log("Creating Osteopath Account")
+          await db.insert(osteopathTable).values({
+            userId: user.id,
+          })
+          console.log("Successfully, Created Osteopathy Account")
+        } catch (error) {
+          console.log(error)
+        }
+      }
       return user;
     };
     const user = await getUser();