+++ src/routes/+layout.server.ts
@@ -1,8 +1,9 @@
 import type { LayoutServerLoad } from './$types';
 
 export const load: LayoutServerLoad = async (event) => {
+
   return {
     logged: event.locals.session?.id ? true : false,
-    user: event.locals.user
+    user: event.locals.user,
   };
 };