+++ src/routes/students/+page.svelte
@@ -8,7 +8,7 @@
 <ul class="flex flex-col gap-y-3">
 	{#each data.users as user}
 		<li class="p-4 shadow flex gap-x-2">
-			<img src={user.image} alt="" />
+			<img width={64} height={64} class="object-contain" src={user.image} alt="" />
 			<div class="flex flex-col">
 				<h3 class="">
 					{user.name}