+++ CHANGELOG.md
@@ -1,13 +1,22 @@
 ## Project Changelog
 
 <a name="x.y.z"></a>
-## x.y.z (2023-09-06)
+
+## 0.0.2
 
 ### Features
-* ...
+
+- Added New TimeSlots Modal
+- Added Osteopaths Carousel
+- Added isOsteopath Guard Function
+- Added Appointment List for Users
 
 ### Bug Fixes
-* ...
+
+- Images size fixed
+-
 
 ### ⚠ BREAKING CHANGES
-* ...
+
+- Removed the Logout Button Instead added a Dropdown for showing options
+- Added Time Availability UI to board